+++ mcp/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add Bevy ECS component metadata to BRP type guide schema information, following Bevy PR [#23102](https://github.com/bevyengine/bevy/pull/23102)
+
 ## [0.20.0-rc.1] - 2026-05-24
 
 ### Breaking Changes